feat(ui): add options form validation
