+++ package.json
@@ -25,7 +25,7 @@
     "@commitlint/cli": "^12.0.0",
     "@commitlint/config-conventional": "^12.0.0",
     "husky": "^6.0.0",
-    "lerna": "^4.0.0",
+    "lerna": "^3.0.0",
     "markdown-toc": "^1.2.0"
   }
 }
